2054: Refactor redux
