+++ accessibility_monitoring_platform/apps/exports/views.py
@@ -22,9 +22,9 @@ from ..simplified.utils import (
     record_simplified_model_create_event,
     record_simplified_model_update_event,
 )
-from .csv_export_utils import download_equality_body_cases
 from .forms import ExportConfirmForm, ExportCreateForm, ExportDeleteForm
 from .models import Export, ExportCase
+from .utils import download_equality_body_cases
 
 
 class EnforcementBodyMixin:
